refactor(core,cli): move the goal guard into @cline/core for cross-surface reuse
The goal state machine, mark_goal_complete tool, verification prompt
formatting/detection, and the awaited turn wrapper move from apps/cli to
sdk/packages/core/src/session/goal-guard.ts (tests move with them); the CLI
re-exports everything unchanged from its existing module path. This lets
the VS Code extension share the exact same guard and prompt strings, so
transcripts written by one surface hydrate correctly in the other.
